configs: am62x_evm_a53: Remove inclusion of USB DFU config
Currently, USB DFU for the A53 SPL and U-Boot are enabled by default. This
is because the USB DFU config fragment, am62x_a53_usbdfu.config, is
included in am62's A53 defconfig. Note that the R5 SPL defconfig does not
include the corresponding am62x_r5_usbdfu.config so it is not built by
default with USB DFU enabled. The user must add the config fragment
themselves.

However, if USB DFU is enabled using these config fragments, the serial
flash DFU backend does not work, which for example is used by the EFI
capsule update mechanism. This is because the current implementation of the
SF DFU backend requires that the DFU data buffer size be the same as the
sector erase size, and the USB DFU config fragment (for both A53 and R5)
reduces the size of the DFU data buffer so that it can fit into SPL's much
smaller heap, which then it is not the same as the sector erase size.

This patch removes the default inclusion of the A53 USB DFU config
fragment. This allows SF DFU to work in the default build. If the user
wants to enable DFU boot, they must include both the A53 and R5 USB DFU
boot config fragments.

Fixes: dfc2dff5a84 ("configs: am62x_evm_*: Enable USB and DFU support")

Signed-off-by: Jonathan Humphreys <j-humphreys@ti.com>